remove portsforobject from factory
