+++ pkg/kubectl/polymorphichelpers/BUILD
@@ -4,10 +4,12 @@ go_library(
     name = "go_default_library",
     srcs = [
         "attachablepodforobject.go",
+        "canbeautoscaled.go",
         "helpers.go",
         "historyviewer.go",
         "interface.go",
         "logsforobject.go",
+        "portsforobject.go",
         "statusviewer.go",
         "updatepodspec.go",
     ],
@@ -36,6 +38,7 @@ go_library(
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/labels:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes:go_default_library",
         "//vendor/k8s.io/client-go/rest:go_default_library",
@@ -47,6 +50,7 @@ go_test(
     srcs = [
         "helpers_test.go",
         "logsforobject_test.go",
+        "portsforobject_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
@@ -63,6 +67,7 @@ go_test(
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/diff:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
         "//vendor/k8s.io/client-go/testing:go_default_library",
     ],
